+++ packages/tizen_app_manager/CHANGELOG.md
@@ -1,4 +1,15 @@
+## 0.1.2
+
+* Implement a Dart finalizer for `AppRunningContext`.
+* Deprecate `AppRunningContext.dispose`.
+* Clean up README.
+
+## 0.1.1
+
+* Add the main exporter file `tizen_app_manager.dart`.
+* Add an optional parameter `background` to `AppRunningContext.terminate()`
+  to support terminating background applications.
+
 ## 0.1.0
 
 * Initial release.
-